+++ history.md
@@ -77,6 +77,8 @@ This release is largely API compatible, but makes several breaking changes.
 - Refactor URI parsing to happen earlier, in Request initialization.
 - When adding URL params, handle URLs that already contain params.
 - Add a few more exception classes for obscure HTTP status codes.
+- Multipart: use a much more robust multipart boundary with greater entropy.
+- Make `RestClient::Payload::Base#inspect` stop pretending to be a String.
 
 # 2.0.0.rc1